requirements.txt
absl-py==1.3.0
asgiref==3.5.2
astunparse==1.6.3
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.3
django-admin==2.0.1
django-cors-headers==3.13.0
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-six==1.0.4
djangorestframework==3.14.0
flatbuffers==22.10.26
gast==0.4.0
google-auth==2.14.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.50.0
h5py==3.7.0
idna==3.4
keras==2.10.0
Keras-Preprocessing==1.1.2
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
numpy==1.23.4
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==21.3
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
pytz==2022.6
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
screen==1.0.1
six==1.16.0
sqlparse==0.4.3
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.1.0
typing_extensions==4.4.0
tzdata==2022.6
urllib3==1.26.12
Werkzeug==2.2.2
wrapt==1.14.1
xlwt==1.3.0
